Created a few scripts specific to the proteasome.
One is for plotting a datamatrix and coloring based on protein's role in complex (lid, core, other)

The other script is for diagonsing score function performance.
The example is of a bicluster made of the proteasome lid and scored
with a few different score functions.  Also compared is the normalization schemes
used to massage input data.