compute_benchmark_probs is a driver script to iterate through the pdb benchmark, call relative stoichiometry function and store results.
analyze_benchmark is an analysis script which compares predictions to true stoichiometries in benchmark

normalize_benchmark_msds is a pre-conditioning script to normalize the data based on length and threshold (experimental).

With this iteration of the code, out of 113 pdbs in benchmark (has greater than 10 data points/ calculated ratios:
score: 3740
top_ten: 15
top_one: 8

score is the sum of all the top ranks in the benchmark. Low is better where rank 0 is top hit was correct.  Average rank is 33. Not great but just a first pass.