Added fuzzy match of input species and queried Uniprot return species to match protein ids sent to get ortholog.
There must be a better way to do this.  The problem is when protein ids
from a species are queried for orthologs in the same species, there is
no database table for speciesA->speciesA.  One way to deal with this is
to create db tables for speciesA->speciesA.  The way it is dealt with
here is to query uniprot for the species of an unmatched protid and
compare that to the input species.  The two are not in the same format
though: Homo Sapiens != hsapiens .  Enter fuzzy matching which is scary
because I had to put some threshold on the fuzzy match and I am not sure
it is a good one.  I can think of a bunch of different ways this could
fail.

Also created functionality to query pdbsws for mapping between pdb chain
id and uniprot ids