Bug with where molecular standards were placed, removed from plot.
There is a weird bug within matplotlib when creating a new axis using
twinx() that when you change the tick marks, the plot shifts over and so
does the tick marks.  I did not fully characterize the bug but stay
skeptical of twinx.

Also, added fitting of peaks with gaussian.  Some of it is hard coded to
commander complex and probably does not make total sense for all
complexes.