Script to test non-negative matrix factorization.
The script reads in a msds pickle and set of proteins.
The data matrix is then run through nimfa's nonnegative matrix
factorization.  A rank can be specified on the commandline which defines
how many clusters.  Each cluster is converted to a bicluster object and
scored.  A threshold can be specified as well to alter what
proteins/fractions get added to the bicluster.

Testing on proteasome. Scoring with zscore_all.  The main cluster
(all proteasome subunits) gets a high score but the lid
cluster does not score well possibly due to added fractions that
produced little and bring down the score.  Using these biclusters as
seeds into monte carlo might be a good option.