Added functionality to exclude complexes from precision recall evaluation.
We took a conservative approach when building the training
and test sets by randomly splitting the benchmark and then
removing complexes from the training set that had any
edge in the test set.  We then implemented the clique precision recall
by calling cliques from complexes in the test set true positives and
all other cliques false positives (ie. cliques that spanned multiple complexes in the test set)

This led to false positives for cliques that spanned multiple complexes
in the test set but where from a single complex that may have been
removed during our conservative benchmark creation.  We therefore do not
count cliques to be false positives if they are in the entire gold
standard (passed in by --exclude_complexes commandline parameter).

Another way to think about this is our training and test sets are
sets of cliques and not complexes.  We enumerate all possible cliques from
the space of proteins in the benchmark  and label them true positives or
false positives.  This is infeasible to implement but what the
--exclude_complexes functionality does is to remove the false positives
that were assigned to the training set but erroneously calculated in the
test set by our implementation.

Overall however this does not effect the global well sampled trend of
performance and should effect most prediction sets more or less
equivalently.

See evernote:
https://www.evernote.com/shard/s412/nl/2147483647/01d3585c-069c-4469-8fc4-78629fb83272/