Functionality to trim subunits from predicted complexes that do not have any edges that score higher than a given threshold.
Noticed in the false positives of clique precision recall analysis,
certain subunits really had no connection to the complex, having low
edge weights to all other members of the complex.  Seemed to be an issue
with MCL and not Neuman method.  Added a script and an extra processing
step to clustering optimization to remove subunits that do not have
edges above a given threshold.  The threshold is generally defined by
the fraction parameter in clustering optimization.