Fixed bug with calculating unique items.
Not sure how this ever ran without throwing an error because the unit
test fails with an error. The bug was introduced in commit:
a1321c4025c2f349e3fa82c6e7586ff17ad3630e . It was due to setting
'bait_id_column_str' as the index and then trying to access
'bait_id_column_str'. This is stupid anyway because I should be able to
access the index column by its name but pandas doesn't allow that for
some reason.  Anyway, the solution is to grab the df.index .